feat(aggregate): update AVG function to return original column type
Modified the AVG aggregation function to return results in the original column type instead of always returning DOUBLE. This change enhances type consistency across aggregate operations. Additionally, updated related tests to reflect the new behavior, ensuring accurate type assertions for average calculations.

+++ integration/csharp/tests/Duckstax.Otterbrix.Tests/WrapperTest.cs
@@ -167,7 +167,7 @@ public class Tests
                 Assert.IsTrue(nameVal.GetString() == "Name " + number.ToString());
                 Assert.IsTrue(countVal.GetInt() == 10);
                 Assert.IsTrue(sumVal.GetInt() == 5 * (number % 20) + 5 * ((number + 10) % 20));
-                Assert.IsTrue(avgVal.GetDouble() == (number % 20 + (number + 10) % 20) / 2);
+                Assert.IsTrue(avgVal.GetInt() == (number % 20 + (number + 10) % 20) / 2);
                 Assert.IsTrue(minVal.GetInt() == number % 20);
                 Assert.IsTrue(maxVal.GetInt() == (number + 10) % 20);
             }
@@ -191,7 +191,7 @@ public class Tests
                 Assert.IsTrue(nameVal.GetString() == "Name " + number.ToString());
                 Assert.IsTrue(countVal.GetInt() == 10);
                 Assert.IsTrue(sumVal.GetInt() == 5 * (number % 20) + 5 * ((number + 10) % 20));
-                Assert.IsTrue(avgVal.GetDouble() == (number % 20 + (number + 10) % 20) / 2);
+                Assert.IsTrue(avgVal.GetInt() == (number % 20 + (number + 10) % 20) / 2);
                 Assert.IsTrue(minVal.GetInt() == number % 20);
                 Assert.IsTrue(maxVal.GetInt() == (number + 10) % 20);
             }
